+++ tests/test_checker/test_hypothesis.py
@@ -1,0 +1,74 @@
+# -*- coding: utf-8 -*-
+
+"""
+Integration test that our linter does not brake on different random programs.
+
+We generate thousands of them with the help of ``hypothesis`` and test
+that they are parsed and processed correctly.
+
+We cannot tell whether or not violations are correctly shown by a random input,
+but we can tell that our program did not raise any exception at least.
+
+See also:
+    https://github.com/HypothesisWorks/hypothesis
+    https://github.com/Zac-HD/hypothesmith
+
+"""
+
+import io
+import tokenize
+
+import hypothesmith
+from hypothesis import HealthCheck, given, reject, settings
+
+from wemake_python_styleguide.checker import Checker
+
+pytest_plugins = [
+    'plugins.ast_tree',
+    'plugins.tokenize_parser',
+]
+
+settings.register_profile(
+    'slow', deadline=None, suppress_health_check=HealthCheck.all(),
+)
+settings.load_profile('slow')
+
+
+def _fixup(string: str) -> str:
+    """Avoid known issues with tokenize() by editing the string."""
+    return ''.join(
+        char for char in string
+        if char.isprintable()
+    ).strip().strip('\\').strip() + '\n'
+
+
+@given(source_code=hypothesmith.from_grammar().map(_fixup))
+def test_no_exceptions(
+    source_code,
+    default_options,
+    parse_ast_tree,
+    parse_tokens,
+):
+    """
+    This testcase is a complex example of magic.
+
+    We use property based-test to generate python programs for us.
+    And then we ensure that our linter does not crash on arbitary input.
+    """
+    try:
+        tree = parse_ast_tree(str(source_code.encode('utf-8-sig')))
+    except (UnicodeEncodeError, SyntaxError):
+        reject()
+        raise
+
+    lines = io.StringIO(source_code)
+    tokens = list(tokenize.generate_tokens(lambda: next(lines)))
+
+    Checker.parse_options(default_options)
+    checker = Checker(tree, tokens)
+
+    for violation in checker.run():
+        assert isinstance(violation[0], int)
+        assert isinstance(violation[1], int)
+        assert 'WPS000' not in violation[2]
+        assert violation[3] == Checker